+++ internal/repos/syncer_test.go
@@ -939,7 +939,7 @@ func testSyncerMultipleServices(store repos.Store) func(t *testing.T) {
 
 		syncer := &repos.Syncer{
 			Logger: logtest.Scoped(t),
-			Sourcer: func(service *types.ExternalService) (repos.Source, error) {
+			Sourcer: func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 				s, ok := sourcers[service.ID]
 				if !ok {
 					t.Fatalf("sourcer not found: %d", service.ID)
@@ -1069,7 +1069,7 @@ func testOrphanedRepo(store repos.Store) func(*testing.T) {
 		syncer := &repos.Syncer{
 
 			Logger: logtest.Scoped(t),
-			Sourcer: func(service *types.ExternalService) (repos.Source, error) {
+			Sourcer: func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 				s := repos.NewFakeSource(svc1, nil, githubRepo)
 				return s, nil
 			},
@@ -1081,7 +1081,7 @@ func testOrphanedRepo(store repos.Store) func(*testing.T) {
 		}
 
 		// Sync second service
-		syncer.Sourcer = func(service *types.ExternalService) (repos.Source, error) {
+		syncer.Sourcer = func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 			s := repos.NewFakeSource(svc2, nil, githubRepo)
 			return s, nil
 		}
@@ -1096,7 +1096,7 @@ func testOrphanedRepo(store repos.Store) func(*testing.T) {
 		assertDeletedRepoCount(ctx, t, store, 0)
 
 		// Remove the repo from one service and sync again
-		syncer.Sourcer = func(services *types.ExternalService) (repos.Source, error) {
+		syncer.Sourcer = func(ctx context.Context, services *types.ExternalService) (repos.Source, error) {
 			s := repos.NewFakeSource(svc1, nil)
 			return s, nil
 		}
@@ -1120,7 +1120,7 @@ func testOrphanedRepo(store repos.Store) func(*testing.T) {
 		assertDeletedRepoCount(ctx, t, store, 0)
 
 		// Remove the repo from the second service and sync again
-		syncer.Sourcer = func(service *types.ExternalService) (repos.Source, error) {
+		syncer.Sourcer = func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 			s := repos.NewFakeSource(svc2, nil)
 			return s, nil
 		}
@@ -1170,7 +1170,7 @@ func testCloudDefaultExternalServicesDontSync(store repos.Store) func(*testing.T
 		logger := logtest.Scoped(t)
 		syncer := &repos.Syncer{
 			Logger: logger,
-			Sourcer: func(service *types.ExternalService) (repos.Source, error) {
+			Sourcer: func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 				s := repos.NewFakeSource(svc1, nil, githubRepo)
 				return s, nil
 			},
@@ -1229,7 +1229,7 @@ func testConflictingSyncers(store repos.Store) func(*testing.T) {
 		// Sync first service
 		syncer := &repos.Syncer{
 			Logger: logtest.Scoped(t),
-			Sourcer: func(service *types.ExternalService) (repos.Source, error) {
+			Sourcer: func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 				s := repos.NewFakeSource(svc1, nil, githubRepo)
 				return s, nil
 			},
@@ -1243,7 +1243,7 @@ func testConflictingSyncers(store repos.Store) func(*testing.T) {
 		// Sync second service
 		syncer = &repos.Syncer{
 			Logger: logtest.Scoped(t),
-			Sourcer: func(service *types.ExternalService) (repos.Source, error) {
+			Sourcer: func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 				s := repos.NewFakeSource(svc2, nil, githubRepo)
 				return s, nil
 			},
@@ -1288,7 +1288,7 @@ func testConflictingSyncers(store repos.Store) func(*testing.T) {
 		// Start syncing using tx1
 		syncer = &repos.Syncer{
 			Logger: logtest.Scoped(t),
-			Sourcer: func(service *types.ExternalService) (repos.Source, error) {
+			Sourcer: func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 				s := repos.NewFakeSource(svc1, nil, updatedRepo)
 				return s, nil
 			},
@@ -1301,7 +1301,7 @@ func testConflictingSyncers(store repos.Store) func(*testing.T) {
 
 		syncer2 := &repos.Syncer{
 			Logger: logtest.Scoped(t),
-			Sourcer: func(service *types.ExternalService) (repos.Source, error) {
+			Sourcer: func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 				s := repos.NewFakeSource(svc2, nil, githubRepo.With(func(r *types.Repo) {
 					r.Description = newDescription
 				}))
@@ -1387,7 +1387,7 @@ func testSyncRepoMaintainsOtherSources(store repos.Store) func(*testing.T) {
 		// Sync first service
 		syncer := &repos.Syncer{
 			Logger: logtest.Scoped(t),
-			Sourcer: func(service *types.ExternalService) (repos.Source, error) {
+			Sourcer: func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 				s := repos.NewFakeSource(svc1, nil, githubRepo)
 				return s, nil
 			},
@@ -1401,7 +1401,7 @@ func testSyncRepoMaintainsOtherSources(store repos.Store) func(*testing.T) {
 		// Sync second service
 		syncer = &repos.Syncer{
 			Logger: logtest.Scoped(t),
-			Sourcer: func(service *types.ExternalService) (repos.Source, error) {
+			Sourcer: func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 				s := repos.NewFakeSource(svc2, nil, githubRepo)
 				return s, nil
 			},
@@ -1504,7 +1504,7 @@ func testUserAddedRepos(store repos.Store) func(*testing.T) {
 		// Admin service will sync both repos
 		syncer := &repos.Syncer{
 			Logger: logtest.Scoped(t),
-			Sourcer: func(service *types.ExternalService) (repos.Source, error) {
+			Sourcer: func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 				s := repos.NewFakeSource(adminService, nil, publicRepo, privateRepo)
 				return s, nil
 			},
@@ -1521,7 +1521,7 @@ func testUserAddedRepos(store repos.Store) func(*testing.T) {
 		// Unsync the repo to clean things up
 		syncer = &repos.Syncer{
 			Logger: logtest.Scoped(t),
-			Sourcer: func(service *types.ExternalService) (repos.Source, error) {
+			Sourcer: func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 				s := repos.NewFakeSource(adminService, nil)
 				return s, nil
 			},
@@ -1538,7 +1538,7 @@ func testUserAddedRepos(store repos.Store) func(*testing.T) {
 		// By default, user service can only sync public code, even if they have access to private code
 		syncer = &repos.Syncer{
 			Logger: logtest.Scoped(t),
-			Sourcer: func(service *types.ExternalService) (repos.Source, error) {
+			Sourcer: func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 				s := repos.NewFakeSource(userService, nil, publicRepo, privateRepo)
 				return s, nil
 			},
@@ -1561,7 +1561,7 @@ func testUserAddedRepos(store repos.Store) func(*testing.T) {
 
 		syncer = &repos.Syncer{
 			Logger: logtest.Scoped(t),
-			Sourcer: func(service *types.ExternalService) (repos.Source, error) {
+			Sourcer: func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 				s := repos.NewFakeSource(userService, nil, publicRepo, privateRepo)
 				return s, nil
 			},
@@ -1589,7 +1589,7 @@ func testUserAddedRepos(store repos.Store) func(*testing.T) {
 
 		syncer = &repos.Syncer{
 			Logger: logtest.Scoped(t),
-			Sourcer: func(service *types.ExternalService) (repos.Source, error) {
+			Sourcer: func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 				s := repos.NewFakeSource(userService, nil, publicRepo, privateRepo)
 				return s, nil
 			},
@@ -1611,7 +1611,7 @@ func testUserAddedRepos(store repos.Store) func(*testing.T) {
 		// Attempt to add some repos with a per user limit set
 		syncer = &repos.Syncer{
 			Logger: logtest.Scoped(t),
-			Sourcer: func(service *types.ExternalService) (repos.Source, error) {
+			Sourcer: func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 				s := repos.NewFakeSource(userService, nil, publicRepo, publicRepo2)
 				return s, nil
 			},
@@ -1626,7 +1626,7 @@ func testUserAddedRepos(store repos.Store) func(*testing.T) {
 		// Attempt to add some repos with a total limit set
 		syncer = &repos.Syncer{
 			Logger: logtest.Scoped(t),
-			Sourcer: func(service *types.ExternalService) (repos.Source, error) {
+			Sourcer: func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 				s := repos.NewFakeSource(userService, nil, publicRepo, publicRepo2)
 				return s, nil
 			},
@@ -1695,7 +1695,7 @@ func testNameOnConflictOnRename(store repos.Store) func(*testing.T) {
 		// Sync first service
 		syncer := &repos.Syncer{
 			Logger: logtest.Scoped(t),
-			Sourcer: func(service *types.ExternalService) (repos.Source, error) {
+			Sourcer: func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 				s := repos.NewFakeSource(svc1, nil, githubRepo1)
 				return s, nil
 			},
@@ -1709,7 +1709,7 @@ func testNameOnConflictOnRename(store repos.Store) func(*testing.T) {
 		// Sync second service
 		syncer = &repos.Syncer{
 			Logger: logtest.Scoped(t),
-			Sourcer: func(*types.ExternalService) (repos.Source, error) {
+			Sourcer: func(context.Context, *types.ExternalService) (repos.Source, error) {
 				s := repos.NewFakeSource(svc2, nil, githubRepo2)
 				return s, nil
 			},
@@ -1728,7 +1728,7 @@ func testNameOnConflictOnRename(store repos.Store) func(*testing.T) {
 		// Sync first service
 		syncer = &repos.Syncer{
 			Logger: logtest.Scoped(t),
-			Sourcer: func(*types.ExternalService) (repos.Source, error) {
+			Sourcer: func(context.Context, *types.ExternalService) (repos.Source, error) {
 				s := repos.NewFakeSource(svc1, nil, renamedRepo1)
 				return s, nil
 			},
@@ -1802,7 +1802,7 @@ func testDeleteExternalService(store repos.Store) func(*testing.T) {
 		// Sync first service
 		syncer := &repos.Syncer{
 			Logger: logtest.Scoped(t),
-			Sourcer: func(service *types.ExternalService) (repos.Source, error) {
+			Sourcer: func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 				s := repos.NewFakeSource(svc1, nil, githubRepo)
 				return s, nil
 			},
@@ -1816,7 +1816,7 @@ func testDeleteExternalService(store repos.Store) func(*testing.T) {
 		// Sync second service
 		syncer = &repos.Syncer{
 			Logger: logtest.Scoped(t),
-			Sourcer: func(service *types.ExternalService) (repos.Source, error) {
+			Sourcer: func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 				s := repos.NewFakeSource(svc2, nil, githubRepo)
 				return s, nil
 			},
@@ -1929,7 +1929,7 @@ func testAbortSyncWhenThereIsRepoLimitError(store repos.Store) func(*testing.T) 
 			// Sync first service
 			syncer := &repos.Syncer{
 				Logger: logger,
-				Sourcer: func(service *types.ExternalService) (repos.Source, error) {
+				Sourcer: func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 					s := repos.NewFakeSource(svc, nil, githubRepo, githubRepo2)
 					return s, nil
 				},
@@ -2067,7 +2067,7 @@ func testUserAndOrgReposAreCountedCorrectly(store repos.Store) func(*testing.T) 
 			// Sync first service
 			syncer := &repos.Syncer{
 				Logger: logger,
-				Sourcer: func(service *types.ExternalService) (repos.Source, error) {
+				Sourcer: func(ctx context.Context, service *types.ExternalService) (repos.Source, error) {
 					s := repos.NewFakeSource(svc, nil, reposToSync[repoIdx], reposToSync[repoIdx+1])
 					return s, nil
 				},
